fix(ext4): fast-commit AddRange parses on-disk ext4_extent (unwritten + 48-bit phys)
replay_fast_commit mis-parsed the ext4_fc_add_range record. After fc_ino,
the payload is a 12-byte on-disk ext4_extent (ee_block:le32, ee_len:le16,
ee_start_hi:le16, ee_start_lo:le32), but the parser read len as a full u32
at offset 8 (folding ee_len with ee_start_hi) and physical_block as a u32 at
offset 12 (dropping the high 16 physical bits). It also ignored ee_len's
unwritten-extent encoding (ee_len > 32768 => unwritten, real len = ee_len -
32768).

Decode exactly as the kernel does (ext4_ext_get_actual_len / ext4_ext_pblock
/ ext4_ext_is_unwritten): FcExtentRange.physical_block is now u64 (48-bit) and
gains an unwritten flag. The old replay_add_range test had encoded the wrong
4xu32 layout; rebuilt it on the real ext4_extent layout and added
replay_add_range_decodes_unwritten_and_48bit_physical_bd_6nwjx (unwritten
10-block extent at physical 0x3_0000_5000) which fails on the old parser.

Makes the FC parser kernel-correct ahead of FC-apply (bd-6nwjx). ffs-journal
245/245, clippy clean; ffs-core compiles (verify signature widened to u64).

+++ crates/ffs-core/src/lib.rs
@@ -4121,7 +4121,7 @@ impl OpenFs {
         ino: u32,
         logical_block: u32,
         len: u32,
-        physical_block: Option<u32>,
+        physical_block: Option<u64>,
         op: &'static str,
     ) -> bool {
         match self.read_inode(cx, InodeNumber(u64::from(ino))) {
